Disgenet_Extraction
This Python script enables users to search for disease variants based on a disease UMLS CUI ID. It utilizes web scraping techniques to extract data from the DisGeNET database and exports the retrieved information to a CSV file. Users can specify the number of data to fetch and export according to their requirements.